+++ src/hooks/useMaxCharacters.ts
@@ -11,8 +11,8 @@ export const useMaxCharacters = (
             if (isOpen) {
                 setText(children);
             } else {
-                setText(children.trim().substring(0, maxCharacters));
+                setText(children.substring(0, maxCharacters));
             }
         }
-    }, [maxCharacters, isOpen]);
+    }, [maxCharacters, isOpen, children]);
 };