+++ src/hooks/useMaxWords.ts
@@ -11,11 +11,8 @@ export const useMaxWords = (
             if (isOpen) {
                 setText(children);
             } else {
-                const wordsArray = children
-                    .trim()
-                    .split(' ')
-                    .filter((c) => c !== '');
-                setText(wordsArray.slice(0, maxWords).join(' ').trim());
+                const wordsArray = children.split(' ').filter((c) => c !== '');
+                setText(wordsArray.slice(0, maxWords).join(' '));
             }
         }
     }, [maxWords, isOpen, children]);