+++ src/utils/index.ts
@@ -1,13 +1,7 @@
 export const isAllText = (truncatedText, text) => {
     return (
         truncatedText &&
-        truncatedText
-            .trim()
-            .split('')
-            .filter((c) => c !== ' ').length ===
-            text
-                .trim()
-                .split('')
-                .filter((c) => c !== ' ').length
+        truncatedText.split('').filter((c) => c !== ' ').length ===
+            text.split('').filter((c) => c !== ' ').length
     );
 };